fix(sprint/bucket-exchange/1): Nolan RF-1·RF-2·RF-3 수정
RF-1 (Blocker): POST /api/applications — runTransaction으로 중복 확인·삽입 원자화
  TOCTOU race condition: where().get() → .add() 사이 동시 요청 시 중복 가능
  → tx.get(duplicateQuery) + tx.set() + tx.update()를 단일 트랜잭션으로 묶음

RF-2 (Major): firestore.rules 추가
  quests 공개 읽기, applications 본인만 읽기, 클라이언트 직접 쓰기 전체 차단

RF-3 (Minor): 타입 단일 소스 정리
  QuestStatus·QuestCategory·Difficulty·Quest·Applicant → types/quest.ts로 이전
  data/quests.ts는 목 데이터·헬퍼만 유지, types/quest.ts에서 임포트
  FirestoreQuest도 types/quest.ts로 통합, data/quests.ts 의존 제거
