+++ apps/bucket-exchange/src/app/api/applications/route.ts
@@ -8,7 +8,6 @@ export async function POST(req: NextRequest) {
     if (!token) return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
 
     const decoded = await adminAuth.verifyIdToken(token);
-
     const body = await req.json();
     const { questId, nickname, reason, plan, schedule, desiredReward } = body;
 
@@ -16,41 +15,44 @@ export async function POST(req: NextRequest) {
       return NextResponse.json({ error: 'Missing required fields' }, { status: 400 });
     }
 
-    // 중복 지원 방지
-    const existing = await adminDb
-      .collection('quests')
-      .doc(questId)
-      .collection('applications')
-      .where('applicantId', '==', decoded.uid)
-      .limit(1)
-      .get();
+    const questRef = adminDb.collection('quests').doc(questId);
+    const applicationsRef = questRef.collection('applications');
+    const duplicateQuery = applicationsRef.where('applicantId', '==', decoded.uid).limit(1);
 
-    if (!existing.empty) {
-      return NextResponse.json({ error: 'Already applied' }, { status: 409 });
+    let newId: string | undefined;
+
+    try {
+      await adminDb.runTransaction(async (tx) => {
+        // 중복 확인과 삽입을 단일 트랜잭션으로 묶어 TOCTOU race condition 방지.
+        // tx.get() 이후 다른 요청이 끼어들면 Firestore가 트랜잭션을 retry하거나 abort하므로
+        // 동일 uid의 중복 지원이 원자적으로 차단된다.
+        const existing = await tx.get(duplicateQuery);
+        if (!existing.empty) {
+          throw Object.assign(new Error('Already applied'), { code: 'ALREADY_APPLIED' });
+        }
+
+        const newRef = applicationsRef.doc();
+        newId = newRef.id;
+        tx.set(newRef, {
+          questId,
+          nickname,
+          reason,
+          plan,
+          schedule: schedule ?? null,
+          desiredReward: Number(desiredReward ?? 0),
+          applicantId: decoded.uid,
+          appliedAt: FieldValue.serverTimestamp(),
+        });
+        tx.update(questRef, { applicantCount: FieldValue.increment(1) });
+      });
+    } catch (txErr) {
+      if (txErr instanceof Error && (txErr as Error & { code?: string }).code === 'ALREADY_APPLIED') {
+        return NextResponse.json({ error: 'Already applied' }, { status: 409 });
+      }
+      throw txErr;
     }
 
-    const docRef = await adminDb
-      .collection('quests')
-      .doc(questId)
-      .collection('applications')
-      .add({
-        questId,
-        nickname,
-        reason,
-        plan,
-        schedule: schedule ?? null,
-        desiredReward: Number(desiredReward ?? 0),
-        applicantId: decoded.uid,
-        appliedAt: FieldValue.serverTimestamp(),
-      });
-
-    // applicantCount 증가
-    await adminDb
-      .collection('quests')
-      .doc(questId)
-      .update({ applicantCount: FieldValue.increment(1) });
-
-    return NextResponse.json({ id: docRef.id }, { status: 201 });
+    return NextResponse.json({ id: newId }, { status: 201 });
   } catch (err) {
     console.error('POST /api/applications', err);
     return NextResponse.json({ error: 'Failed to submit application' }, { status: 500 });
